% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_freq_mean_se.r
\name{CalcFreqMeanSE}
\alias{CalcFreqMeanSE}
\title{Calculates mean and standard error of frequency data}
\usage{
CalcFreqMeanSE(bins, N, frdist)
}
\arguments{
\item{bins}{the values in the class marks (ages in this case)}

\item{N}{the total number (ESS in this case)}

\item{frdist}{the frequency distribution (should sum to one over the bins)}
}
\value{
vector of mean and standard error
}
\description{
Helper function for Francis ESS calculations.
}
