% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_survey_to_at_age.r
\name{ConvertSurveyToAtAge}
\alias{ConvertSurveyToAtAge}
\title{Converts surveys to age matrices}
\usage{
ConvertSurveyToAtAge(asap)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}
}
\value{
list with catch at age matrices by index
}
\description{
Converts West Coast style surveys (total abundance and age comps) into catch at age matrices for age consistency plots.
}
