% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_docxASAP_control_file.r
\name{createDefaultdocxASAPcontrolfile}
\alias{createDefaultdocxASAPcontrolfile}
\title{createDefaultdocxASAPcontrolfile}
\usage{
createDefaultdocxASAPcontrolfile(wd)
}
\arguments{
\item{wd}{directory where ASAP run is located}
}
\description{
Create docxASAP.csv with default settings.
}
