% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_param_list.R
\name{CreateParamList}
\alias{CreateParamList}
\title{CreateParamList}
\usage{
CreateParamList(asap.pin.obj, asap.dat.obj)
}
\arguments{
\item{asap.pin.obj}{contains data and comments from an ASAP3 pin file}

\item{asap.dat.obj}{contains data and comments from an ASAP3 dat file}
}
\description{
Provides the upper and lower limits for each parameter based on ASAP tpl bounded parameter values.
}
