% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_ASAP.r
\name{docxASAP}
\alias{docxASAP}
\title{docxASAP}
\usage{
docxASAP(wd, asap.name, od = NULL, docx.name = "docxASAP.docx",
  control.file = "docxASAP.csv", first.figure.number = 1,
  figure.prefix = "", plotf = "png", use.group = NULL,
  append.asap.name.caption = TRUE)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{od}{output directory for plots and csv files (default = NULL means wd\\plots\\)}

\item{docx.name}{name of Word document to create (default = docxASAP.docx)}

\item{control.file}{csv file of parameters defining plots to add to docx (default = docxASAP.csv)}

\item{first.figure.number}{starting value for figures, succesive figures based on Order}

\item{figure.prefix}{text to add before figure number, e.g., "B" becomes Figure B1. (default = "")}

\item{plotf}{type of plot to save (default = 'png')}

\item{use.group}{only add figures from this Group to docx (default = NULL means all groups added)}

\item{append.asap.name.caption}{flag to add (TRUE = default) asap.name to end of figure caption text}
}
\description{
Create a Word document of the figures from ASAPplots with predefined captions.
}
