% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixed_params.R
\name{GetFixedParams}
\alias{GetFixedParams}
\title{GetFixedParams}
\usage{
GetFixedParams(asap.dat)
}
\arguments{
\item{asap}{asap dat object from ReadASAP3DatFile}
}
\description{
Returns data frame of parameters for a given ASAP run as either "fixed" or "estimated."
}
