% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_aux_files.r
\name{GrabAuxFiles}
\alias{GrabAuxFiles}
\title{Grab Auxilliary Files}
\usage{
GrabAuxFiles(asap.name, asap, fleet.names, index.names)
}
\arguments{
\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{index.names}{names of indices}
}
\value{
list of summary statistics about run (e.g. # parameters, correlation matrix, max gradient)
}
\description{
This function reads in values from the *.std, *.par, and *.cor files.
}
