% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_dat_file_index_sel_options.r
\name{GrabDatFileIndexSelOptions}
\alias{GrabDatFileIndexSelOptions}
\title{Grab Dat File Index Selectivity Options}
\usage{
GrabDatFileIndexSelOptions(wd, asap.name, asap)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{asap}{name of the variable that read in the asap.rdat file}
}
\value{
list of fleet names and index names
}
\description{
This function gets index selectivity options from the original ASAP.DAT file.
}
