% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_asap.R
\name{JitterASAP}
\alias{JitterASAP}
\title{JitterASAP
Changes .pin file to use random restarts from either full range of parameters or SS3.30 approach of values near estimated ones.}
\usage{
JitterASAP(in.pin, param.list, jitterfac = 0.1)
}
\arguments{
\item{in.pin}{pin object of original parameter estimates}

\item{jitterfac}{factor used in SS3.30 approach (default=0.1)}

\item{.list}{list of parameters as either "fixed" or "estimated", only "estimated" are jittered}
}
\description{
JitterASAP
Changes .pin file to use random restarts from either full range of parameters or SS3.30 approach of values near estimated ones.
}
