% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_biomass_types.r
\name{PlotAllBiomassTypes}
\alias{PlotAllBiomassTypes}
\title{Plot all biomass types}
\usage{
PlotAllBiomassTypes(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Comparison of Total, SSB, and Exploitable Biomass calculated on Jan-1. The Total B is direct from ASAP, while exploitable B and SSB are calculated for this comparison.
}
