% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_SPR_targets.r
\name{PlotAnnualSPRtargets}
\alias{PlotAnnualSPRtargets}
\title{Plot annual SPR targets}
\usage{
PlotAnnualSPRtargets(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Computes and plots SPR targets using selectivity, M, WAA, and maturity by year.
}
