% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ASAP.r
\name{PlotASAP}
\alias{PlotASAP}
\title{Plot ASAP}
\usage{
PlotASAP(wd, asap.name, nyrs.ave = 5, correlation.limit = 0.9,
  scale.catch.bubble.resid = 2, scale.index.bubble.resid = 2,
  scale.catch.bubble.data = 6, scale.index.bubble.data = 6,
  mcmc.burn = 0, mcmc.thin = 1, first.age = -999,
  save.plots = TRUE, plotf = "png", make.one.pdf = TRUE)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{nyrs.ave}{number of years to average for calculating Reference Points (defaults to 5)}

\item{correlation.limit}{will return parameter combos where |correlation| > value (defaults to 0.9)}

\item{scale.catch.bubble.resid}{larger values increase size of catch age comp bubbles (defaults to 2)}

\item{scale.index.bubble.resid}{larger values increase size of catch age comp bubbles (defaults to 2)}

\item{scale.catch.bubble.data}{larger values increase size of catch age comp bubbles (defaults to 6)}

\item{scale.index.bubble.data}{larger values increase size of catch age comp bubbles (defaults to 6)}

\item{mcmc.burn}{number of realizations to remove from start of MCMC (defaults to 0)}

\item{mcmc.thin}{thinning parameter for MCMC (defaults to 1)}

\item{first.age}{youngest age to use in catch curves, -999 finds peak age (defaults to -999)}

\item{save.plots}{saves indivdual plots (defaults to TRUE)}

\item{plotf}{format for individual plots (defaults to 'png')}

\item{make.one.pdf}{diagnostics/results/ref pts/MCMC/data are all saved as one pdf (defaults to TRUE)}
}
\description{
This function calls all the other functions to create the desired outputs.
}
\examples{
PlotASAP("C:\\\\main.dir\\\\my.asap.results.dir","simple") 
}
