% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catchability.r
\name{PlotCatchability}
\alias{PlotCatchability}
\title{Plot survey catchability}
\usage{
PlotCatchability(asap, index.names, a1, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{a1}{list file produced by grab.aux.files function}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Barplots of survey q with approximate confidence intervals and time series plots of q values (if variable).
}
