% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_age_comp.r
\name{PlotCatchAgeComp}
\alias{PlotCatchAgeComp}
\title{Plot age composition by fleet}
\usage{
PlotCatchAgeComp(asap, fleet.names, save.plots, od, plotf, liz.palette,
  is.catch.flag = TRUE)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}

\item{is.catch.flag}{true means only catch plotted, false also plots discards (defaults to TRUE)}
}
\description{
Compare observed and predicted age composition by year.
}
