% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_age_comp_bubbles.r
\name{PlotCatchAgeCompBubbles}
\alias{PlotCatchAgeCompBubbles}
\title{Bubble plots of catch age comps}
\usage{
PlotCatchAgeCompBubbles(asap, fleet.names, save.plots, od, plotf,
  scale.catch.bubble.data = 6, is.catch.flag = T)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{scale.catch.bubble.data}{larger values increase size of catch age comp bubbles (defaults to 6)}

\item{is.catch.flag}{true means only catch plotted, false also plots discards (defaults to TRUE)}
}
\description{
Makes bubble plots of fishery catch age composition by fleet.
}
