% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_age_comp_resids.r
\name{PlotCatchAgeCompResids}
\alias{PlotCatchAgeCompResids}
\title{Bubble plots of catch age comp resids}
\usage{
PlotCatchAgeCompResids(asap, fleet.names, save.plots, od, plotf,
  scale.catch.bubble.resid = 2, is.catch.flag = TRUE)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{scale.catch.bubble.resid}{larger values increase size of catch age comp bubbles (defaults to 2)}

\item{is.catch.flag}{true means only catch plotted, false also plots discards (defaults to TRUE)}
}
\description{
Pearson residuals for age composition for all years in bubble plot.
}
