% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_at_age_consistency.r
\name{PlotCatchAtAgeConsistency}
\alias{PlotCatchAtAgeConsistency}
\title{Plot catch at age consistency}
\usage{
PlotCatchAtAgeConsistency(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\value{
list with correlation matrices for input and predicted catch at age matrices for each fleet
}
\description{
Lattice style plot of catch at age X vs age Y with each point a cohort for each fleet. Computed for both input and predicted catch at age.
}
