% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_by_fleet.r
\name{PlotCatchByFleet}
\alias{PlotCatchByFleet}
\title{Plot catch by fleet}
\usage{
PlotCatchByFleet(asap, fleet.names, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Barplot of total catch in weight with colors showing fleets.
}
