% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_catch_curves_for_catch.r
\name{PlotCatchCurvesForCatch}
\alias{PlotCatchCurvesForCatch}
\title{Plot catch curves for catch}
\usage{
PlotCatchCurvesForCatch(asap, save.plots, od, plotf, first.age = -999)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{first.age}{youngest age to use in catch curve, -999 finds peak age (defaults to -999)}
}
\description{
Creates catch curves to estimate Z from catch data. Useful???
}
