% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coh.r
\name{PlotCoh}
\alias{PlotCoh}
\title{Lattice style plot of cohorts by age}
\usage{
PlotCoh(matcoh, save.plots, od, plotf, mytitle = "", mylabels = NA)
}
\arguments{
\item{matcoh}{matrix of values with ages as columns and cohorts as rows}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{mytitle}{title for plot (defaults to blank)}

\item{mylabels}{optional labels to write in boxes on diagonal (defaults to showing "age-X")}
}
\value{
correlation matrix
}
\description{
Each point shows age x vs age y of a cohort. Correlation values also shown. Assumes matrix has ages as columns and that cohorts are in rows.
}
