% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CV.r
\name{PlotCV}
\alias{PlotCV}
\title{Plot CV for F, SSB, and Recruitment}
\usage{
PlotCV(asap, a1, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{a1}{list file produced by grab.aux.files function}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Summary plot of estimated precision over time.
}
