% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discard_4_panel.r
\name{PlotDiscard4Panel}
\alias{PlotDiscard4Panel}
\title{Discard 4 panel plots by fleet}
\usage{
PlotDiscard4Panel(asap, fleet.names, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Shows observed and predicted discards regular and log scale, residuals over time, and resid histogram.
}
