% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expected_spawn.r
\name{PlotExpSpawn}
\alias{PlotExpSpawn}
\title{Plot expected spawnings}
\usage{
PlotExpSpawn(asap, a1, nyrs.ave, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{a1}{list file produced by grab.aux.files function}

\item{nyrs.ave}{number of years to average for calculating Reference Points}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Plot expected number of spawnings as a function of fishing mortality rate.
}
