% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fleet_Fmult.r
\name{PlotFleetFmult}
\alias{PlotFleetFmult}
\title{Plot fleet F multipliers}
\usage{
PlotFleetFmult(asap, fleet.names, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Line plots of Fmult (maximum F at age) by fleet, remember that selectivity can change over time.
}
