% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fleet_Neff.r
\name{PlotFleetNeff}
\alias{PlotFleetNeff}
\title{Plot effective sample size (Neff)}
\usage{
PlotFleetNeff(asap, fleet.names, save.plots, od, plotf, liz.palette,
  is.catch.flag = TRUE)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}

\item{is.catch.flag}{true means only catch plotted, false also plots discards (defaults to TRUE)}
}
\description{
Effective sample size is a measure of how precisely age composition is measured. Plots made by fleet.
}
