% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_francis_fleet.r
\name{PlotFrancisFleet}
\alias{PlotFrancisFleet}
\title{Francis effective sample size plots by fleet}
\usage{
PlotFrancisFleet(asap, a1, fleet.names, save.plots, od, plotf,
  is.catch.flag = T)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{a1}{list file produced by grab.aux.files function}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{is.catch.flag}{true means only catch plotted, false also plots discards (defaults to TRUE)}
}
\description{
Compares input ESS with Francis (2011) estimates. Based on Francis 2011 CJFAS 68: 1124-1138.
}
