% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_francis_index.r
\name{PlotFrancisIndex}
\alias{PlotFrancisIndex}
\title{Francis effective sample size plots by index}
\usage{
PlotFrancisIndex(asap, index.names, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Compares input ESS with Francis (2011) estimates. Based on Francis 2011 CJFAS 68: 1124-1138.
}
