% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_front_page.r
\name{PlotFrontPage}
\alias{PlotFrontPage}
\title{Plot front page}
\usage{
PlotFrontPage(wd, asap.name, asap, a1, save.plots, od, plotf)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{a1}{list file produced by grab.aux.files function}

\item{save.plots}{saves indivdual plots (defaults to TRUE)}

\item{od}{output directory for plots and csv files}

\item{plotf}{format for individual plots (defaults to 'png')}
}
\description{
Summary information about ASAP run and version of ASAPplots used.
}
