% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_high_corr.r
\name{PlotHighCorr}
\alias{PlotHighCorr}
\title{Plot high correlations among parameters}
\usage{
PlotHighCorr(a1, correlation.limit, save.plots, od, plotf)
}
\arguments{
\item{a1}{list file produced by grab.aux.files function}

\item{correlation.limit}{will return parameter combos where |correlation| > value}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Identify parameters with high correlation (as defined by user).
}
