% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_high_CVs.r
\name{PlotHighCVs}
\alias{PlotHighCVs}
\title{Plot parameters with unusually high or low CVs}
\usage{
PlotHighCVs(a1, save.plots, od, plotf)
}
\arguments{
\item{a1}{list file produced by grab.aux.files function}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\value{
matrix of parameters that meet criteria
}
\description{
Identify parameters with CVs < 0.05 or CVs > 0.55.
}
