% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index_age_comp_bubbles.r
\name{PlotIndexAgeCompBubbles}
\alias{PlotIndexAgeCompBubbles}
\title{Bubble plots of index age comps}
\usage{
PlotIndexAgeCompBubbles(
  asap,
  index.names,
  save.plots,
  od,
  plotf,
  scale.index.bubble.data = 6
)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{scale.index.bubble.data}{larger values increase size of index age comp bubbles (defaults to 6)}
}
\description{
Makes bubble plots of age composition by index.
}
