% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index_age_comp_resids.r
\name{PlotIndexAgeCompResids}
\alias{PlotIndexAgeCompResids}
\title{Bubble plots of index age comp resids}
\usage{
PlotIndexAgeCompResids(asap, index.names, save.plots, od, plotf,
  scale.index.bubble.resid = 2, orig.bubble.colors = FALSE)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{scale.index.bubble.resid}{larger values increase size of catch age comp bubbles (defaults to 2)}

\item{orig.bubble.colors}{logical TRUE = red, white, FALSE = red, blue (defaults to FALSE)}
}
\description{
Pearson residuals for age composition for all years in bubble plot.
}
