% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index_at_age_consistency.r
\name{PlotIndexAtAgeConsistency}
\alias{PlotIndexAtAgeConsistency}
\title{Plot index at age consistency}
\usage{
PlotIndexAtAgeConsistency(asap, index.names, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\value{
list with correlation matrices for input and predicted catch at age matrices for each fleet
}
\description{
Lattice style plot of index catch at age X vs age Y with each point a cohort for each fleet. Computed for both input and predicted index catch at age.
}
