% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index_input.r
\name{PlotIndexInput}
\alias{PlotIndexInput}
\title{Plot index input}
\usage{
PlotIndexInput(asap, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Rescale indices to mean 1 and stdev 1 then plot on regular and lognormal scale.
}
