% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_index_selectivities.r
\name{PlotIndexSelectivities}
\alias{PlotIndexSelectivities}
\title{Plot index selectivities}
\usage{
PlotIndexSelectivities(asap, index.names, save.plots, od, plotf,
  liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Line plots of selectivities only for indices with multiple ages selected.
}
