% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_indices_4_panel.r
\name{PlotIndices4Panel}
\alias{PlotIndices4Panel}
\title{Indices 4 panel plots}
\usage{
PlotIndices4Panel(asap, index.names, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Shows observed and predicted indices regular and log scale, residuals over time, and resid histogram.
}
