% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jitter.R
\name{PlotJitter}
\alias{PlotJitter}
\title{PlotJitter}
\usage{
PlotJitter(reslist, save.plots, od, plotf, showtitle, ymaxlimit = NULL)
}
\arguments{
\item{reslist}{result of RunJitter function containing objective function values by realization and original value}

\item{save.plots}{saves indivdual plots}

\item{od}{directory where plot will be saved}

\item{plotf}{format for individual plots}

\item{showtitle}{option to add title for plot providing number of realizations better than original, equal to original, and not converged}

\item{ymaxlimit}{replace any objfxn value greater than this value with this value in plot (default = NULL means no replacement of any values)}
}
\description{
Plot results of jitter analysis, objective function values by realization with line for original value.
}
