% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_M.r
\name{PlotM}
\alias{PlotM}
\title{Plot natural mortality rate}
\usage{
PlotM(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Plot mean natural mortality rate as line, colored dots show annual values (if differ from mean).
}
