% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_main_likelihoods.r
\name{PlotMainLikelihoods}
\alias{PlotMainLikelihoods}
\title{Plot main likelihoods by component}
\usage{
PlotMainLikelihoods(asap.name, asap, save.plots, od, plotf, liz.palette,
  max.grad, npar)
}
\arguments{
\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}

\item{max.grad}{maximum gradient at solution}

\item{npar}{number of parameters}
}
\description{
Not a very useful plot, need to replace this with something more informative and less misleading.
}
