% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maturity.r
\name{PlotMaturity}
\alias{PlotMaturity}
\title{Plot maturity}
\usage{
PlotMaturity(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Line plot of mean maturity at age with colored dots showing annual values (if differ from mean).
}
