% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc.r
\name{PlotMCMC}
\alias{PlotMCMC}
\title{Plot MCMC}
\usage{
PlotMCMC(wd, asap.name, asap, mcmc.burn, mcmc.thin, save.plots, od, plotf)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{mcmc.burn}{number of realizations to remove from start of MCMC (defaults to 0)}

\item{mcmc.thin}{thinning parameter for MCMC (defaults to 1)}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Makes a number of diagnostic plots to evaluate convergence in the Monte Carlo Markov Chain results.
}
