% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NAA.r
\name{PlotNAA}
\alias{PlotNAA}
\title{Plot numbers at age}
\usage{
PlotNAA(asap, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Plots both estimated numbers and proportions as stacked barplot.
}
