% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recr_ssb_yr.r
\name{PlotRecrSSByr}
\alias{PlotRecrSSByr}
\title{Scatterplot of SSB, R with 2-digit year}
\usage{
PlotRecrSSByr(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Standard stock-recruitment plot with years labeled. Also creates R/SSB time series plot.
}
