% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment_devs.r
\name{PlotRecruitmentDevs}
\alias{PlotRecruitmentDevs}
\title{Plot recruitment deviations}
\usage{
PlotRecruitmentDevs(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Shows both line and bar plots of SR residuals.
}
