% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RMSE_95CI.r
\name{PlotRMSE95CI}
\alias{PlotRMSE95CI}
\title{Plot of RMSE values with 95 percent CI cone}
\usage{
PlotRMSE95CI(asap, index.names, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{index.names}{names of indices}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Root mean square error by likelihood component relative to 95 percent CI from N(0,1) distribution.
}
