% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SARC_Recr_SSB.r
\name{PlotSARCrecSSB}
\alias{PlotSARCrecSSB}
\title{SARC style plot of recruitment and SSB}
\usage{
PlotSARCrecSSB(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Recruitment lagged to match SSB (recr years 2:nyears; SSB years 1:(nyears-1)).
}
