% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SPR_table.r
\name{PlotSPRtable}
\alias{PlotSPRtable}
\title{Plot SPR table}
\usage{
PlotSPRtable(asap, nyrs.ave, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{nyrs.ave}{number of years to average for calculating Reference Points}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Plot table of values associated with spawners per recruit and compute a number of per recruit reference points.
}
