% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SR_pred_line.r
\name{PlotSRpredLine}
\alias{PlotSRpredLine}
\title{Standard SR plot with Beverton-Holt curve}
\usage{
PlotSRpredLine(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Scatterplot of SSB, R with underlying stock recruitment relationship (SR fit line ends at (S0, R0)).
}
