% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSB_at_age.r
\name{PlotSSBatAge}
\alias{PlotSSBatAge}
\title{Plot SSB at age}
\usage{
PlotSSBatAge(asap, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Plots both estimated SSB and proportions as stacked barplot.
}
