% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SSB_F_trends.r
\name{PlotSSBFtrend}
\alias{PlotSSBFtrend}
\title{Plot SSB and F time series}
\usage{
PlotSSBFtrend(asap, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Two measures of F plotted (max F and Freport) and plotted separately below SSB time series.
}
