% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waa_matrices.r
\name{PlotWAAmatrices}
\alias{PlotWAAmatrices}
\title{Plot weight at age matrices}
\usage{
PlotWAAmatrices(asap, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}
}
\description{
Line plots of all different weight at age matrices.
}
