% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ASAP3_dat_file.R
\name{ReadASAP3DatFile}
\alias{ReadASAP3DatFile}
\title{ReadASAP3DatFile}
\usage{
ReadASAP3DatFile(datf)
}
\arguments{
\item{datf}{full path and name of dat file (including .dat suffix)}
}
\description{
Read ASAP 3 dat file into R (thanks to Tim Miller for providing original version).
}
