% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ASAP3_pin_file.R
\name{ReadASAP3PinFile}
\alias{ReadASAP3PinFile}
\title{ReadASAP3PinFile}
\usage{
ReadASAP3PinFile(pinf)
}
\arguments{
\item{pinf}{full path and name of pin file (including .pin or .par suffix)}
}
\description{
Read ASAP 3 pin (=par) file into R.
}
