% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jitter.R
\name{RunJitter}
\alias{RunJitter}
\title{RunJitter}
\usage{
RunJitter(wd, asap.name, njitter, ploption = "jitter",
  save.plots = TRUE, od = paste0(wd, "\\\\jitter\\\\"),
  plotf = "png", showtitle = FALSE, copy_std_cor_flag = FALSE)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{njitter}{number of realizations to run with different initial guesses}

\item{ploption}{"full" selects initial guesses from full range of possible values, "jitter" uses SS3.30 approach to select values close to parameter estimates (defaults to "jitter")}

\item{save.plots}{saves indivdual plots (defaults to TRUE)}

\item{od}{location where plot is saved (defaults to jitter subdirectory of wd)}

\item{plotf}{format for individual plots (defaults to 'png')}

\item{showtitle}{option to include title of numeric results on objective function plot (defaults to FALSE)}

\item{copy_std_cor_flag}{option to save .std and .cor files for each jitter run (defaults to FALSE)}
}
\description{
Run ASAP a number of times from different initial guesses.
}
