% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ASAP3_dat_file.R
\name{WriteASAP3DatFile}
\alias{WriteASAP3DatFile}
\title{WriteASAP3DatFile}
\usage{
WriteASAP3DatFile(fname, dat.object, header.text)
}
\arguments{
\item{fname}{full directory and file name to be created (including .dat suffix)}

\item{dat.object}{R object containing all the necessary information}

\item{header.text}{text put run description line in input file}
}
\description{
Function to write ASAP 3 dat file (modified from Tim Miller's text_datwrite.R file).
}
