% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ASAP3_pin_file.R
\name{WriteASAP3PinFile}
\alias{WriteASAP3PinFile}
\title{WriteASAP3PinFile}
\usage{
WriteASAP3PinFile(fname, pin.object)
}
\arguments{
\item{fname}{full directory and file name to be created (including .pin suffix)}

\item{pin.object}{R object containing all the necessary information}
}
\description{
Function to write ASAP 3 pin (=par) file
}
