% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_selectivity_decoder.r
\name{MakeSelectivityDecoder}
\alias{MakeSelectivityDecoder}
\title{Make selectivity decoder}
\usage{
MakeSelectivityDecoder(wd, asap, a1, index.names, od)
}
\arguments{
\item{wd}{directory where ASAP run is located}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{a1}{list file produced by grab.aux.files function}

\item{index.names}{names of indices}

\item{od}{output directory for plots and csv files}
}
\description{
Create csv file decoding selectivity parameters for fleet blocks and indices.
}
