% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_MSY.r
\name{PlotAnnualMSY}
\alias{PlotAnnualMSY}
\title{Plot annual MSY}
\usage{
PlotAnnualMSY(asap, a1, save.plots, od, plotf)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{a1}{list file produced by grab.aux.files function}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Plot values associated with MSY using year-specific inputs.
}
