% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_selectivity_decoder.r,
%   R/plot_fleet_sel_blocks.r
\name{PlotFleetSelBlocks}
\alias{PlotFleetSelBlocks}
\alias{PlotFleetSelBlocks}
\title{Make selectivity decoder}
\usage{
PlotFleetSelBlocks(asap, fleet.names, save.plots, od, plotf, liz.palette)

PlotFleetSelBlocks(asap, fleet.names, save.plots, od, plotf, liz.palette)
}
\arguments{
\item{asap}{name of the variable that read in the asap.rdat file}

\item{fleet.names}{names of fleets}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}

\item{liz.palette}{color definitions}

\item{a1}{list file produced by grab.aux.files function}

\item{index.names}{names of indices}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{od}{output directory for plots and csv files}
}
\description{
Create csv file decoding selectivity parameters for fleet blocks and indices.

Line plots of selectivities by block.
}
