% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcmc.r
\name{PlotMCMC}
\alias{PlotMCMC}
\title{Plot MCMC}
\usage{
PlotMCMC(asap.name, asap, save.plots, od, plotf)
}
\arguments{
\item{asap.name}{Base name of original dat file (without the .dat extension)}

\item{asap}{name of the variable that read in the asap.rdat file}

\item{save.plots}{save individual plots}

\item{od}{output directory for plots and csv files}

\item{plotf}{type of plot to save}
}
\description{
Makes a number of diagnostic plots to evaluate convergence in the Monte Carlo Markov Chain results.
}
